ArchMCP-Ppt/requirements.txt
mcp>=0.9.0
boto3>=1.34.0
python-pptx>=0.6.23
pyyaml>=6.0
pillow>=10.0.0
requests>=2.32.0
